Distribuidos/requirements.txt
fastapi
uvicorn
redis
httpx
pydantic
grpcio
dnspython
requests
grpcio-tools
protobuf
matplotlib
